pyproject.toml
[project]
name = "rising-tides"
version = "0.1.0"
description = "Add your description here"
readme = "README.txt"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
]
